fix(link-cache): always bypass cache for dynamic links
Links created with a `null` address are allowed, as long as they
are also created with a dynamic `source`. In this cases we should
absolutely not cache the resultant link, as there is no other way
to differentiate the calls.

A test has also been added for the previously added `bypassCache`
functionality.