refactor(purge): remove bypassPurge (#9)
While bypassing purge worked in isolated integration tests, it did
not in practice. Bypassing purge simply did not kick off the purge
timer for the newly created link, however if the timer was active
for any other links it would still purge the links which the user
requested not be purged.

This patch takes the alternate solution discussed in defaulting to
not caching receiver links, and instead making this feature an
explicit opt-in for those link/stream types.